what a strange problem, I had to brute-force part 2
the Z3 constraint solver solved part 1, although it took a few
seconds. and the number was small enough that I knew it was
absolutely brute-forceable in much less time.

then for part 2, Z3 hit a brick wall. so I tried to
extrapolate from the part 1 answer to see how the new disc
affected things. I failed, but the website helpfully told
me my answer was too *high*.

this told me that, absolutely for sure, part 2 could be
brute-forced. I implemented it in C++ so I could learn how
to use std::regex and *maybe* so I could try the new C++17
parallel execution stuff, except that the single-threaded
version finishes in milliseconds, so never mind.